- split off the handhelds that seem to be based around gpl162xx technology into their own file from generalplus_gpl_unknown.cpp to generalplus_gpl162xx_lcdtype.cpp - identify unkunsp based on new information
